backend/requirements.txt
flask
flask-cors
python-dotenv
authlib
sqlite3